[fix] PR 리뷰 반영해서 코드 수정
