+++ src/main/java/com/zerobase/accountbook/common/config/fcm/FcmConfig.java
@@ -1,0 +1,14 @@
+package com.zerobase.accountbook.common.config.fcm;
+
+import okhttp3.OkHttpClient;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class FcmConfig {
+
+    @Bean
+    public OkHttpClient okHttpClient() {
+        return new OkHttpClient();
+    }
+}
