+++ src/main/resources/application_prod.yml
@@ -1,0 +1,37 @@
+
+spring:
+  datasource:
+    url: jdbc:mariadb://${MARIADB_HOST}localhost:3306/testdb1
+    username: ${MARIADB_USERNAME}
+    password: ${MARIADB_PASSWORD}
+    driver-class-name: org.mariadb.jdbc.Driver
+
+  jpa:
+    generate-ddl: true
+    hibernate:
+      ddl-auto: update
+    show-sql: true
+
+  cache:
+    type: redis
+  redis:
+    host: localhost
+    port: 6379
+    repositories:
+      enabled: false
+
+  mail:
+    host: smtp.gmail.com
+    port: 587
+    username: busyheyazoo@gmail.com
+    password: ${SMTP_PASSWORD}
+    properties:
+      mail.smtp.auth: true
+      mail.smtp.starttls.enable: true
+
+  logging:
+    level:
+      root: debug
+
+jwt:
+  secret: ${JWT_SECRET}
